requirements.txt
pyyaml
cli2>=2.2
